+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # x.x.x
 
+- feat: add improved example project
 - fix: fix bug where localizations were not resolved correctly due to nested `MaterialApp`s
 
 # 0.3.3